Move Clock and Instant to a more general namespace
So that we can use them for things other than telemetry
