+++ src/EventStore.Core/Telemetry/DurationMetric.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Diagnostics.Metrics;
+using EventStore.Core.Time;
 
 namespace EventStore.Core.Telemetry {
 	public class DurationMetric {
